No allocations on `EncodeTime`, run benchmarks
Use per-instance lazily-allocated buffer to avoid allocations
in `EncodeTime`. We can't use use `e.buf`, as it's used to
encode extLen, and it's smaller than required (9 bytes).
Using buffer on stack doesn't work, as Go doesn't have
a way to figure out if `e.write(b)` escapes or not, so it thinks
it always escapes.

Run benchmarks for `make` (and in Travis CI), and fix benchmark
which was failing probably due to some changes in the codebase.

Benchmark, before:

```
BenchmarkTime-12    	20000000	        94.0 ns/op	       8 B/op	       1 allocs/op
```

after:

```
BenchmarkTime-12    	20000000	        76.7 ns/op	       0 B/op	       0 allocs/op
```

+++ bench_test.go
@@ -338,7 +338,7 @@ func BenchmarkQuery(b *testing.B) {
 		if err != nil {
 			b.Fatal(err)
 		}
-		if values[0].(int8) != 10 {
+		if values[0].(int64) != 10 {
 			b.Fatalf("%v != %d", values[0], 10)
 		}
 	}
